requirements.txt
streamlit
pandas
openai
openpyxl
XlsxWriter
python-docx